feat: Implement PostgreSQL repositories for authorization and user management
- Added PermissionRepository for managing permissions with CRUD operations.
- Introduced ResourceRepository for handling resource-related data.
- Created RolePermissionRepository to manage role-permission associations.
- Developed RoleRepository for role management with pagination and filtering.
- Implemented SeedHistoryRepository to track seed execution history.
- Added UserDeviceRepository for managing user devices and their states.
- Created UserNodeRoleRepository for handling user-node-role associations.
- Developed UserRepository for user management with advanced querying capabilities.
- Introduced UserThemeRepository for managing user themes.
- Established BaseRepository for shared functionality across repositories.

+++ src/FAM.Domain/Locations/Entities/Location.cs
@@ -30,13 +30,13 @@ public class Location : BaseEntity, IHasCreationTime, IHasCreator, IHasModificat
     public User? DeletedBy { get; set; }
 
     public string Name { get; private set; } = string.Empty;
-    public int? CompanyId { get; private set; }
+    public long? CompanyId { get; private set; }
     public string? Code { get; private set; }
-    public int? ParentId { get; private set; }
+    public long? ParentId { get; private set; }
     public string? Type { get; private set; }
     public string? FullPath { get; private set; }
     public string? PathIds { get; private set; }
-    public int? CountryId { get; private set; }
+    public long? CountryId { get; private set; }
     public string? Description { get; private set; }
 
     // Navigation properties
